requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
requests>=2.31.0
python-multipart>=0.0.9
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.33.0
safetensors>=0.4.3
diffusers>=0.30.0
transformers>=4.44.0
sentencepiece>=0.2.0
protobuf>=4.25.3
opencv-python>=4.9.0
Pillow>=10.2.0
realesrgan>=0.3.0
basicsr>=1.4.2